#140-create module for NoArrowsSessionScreen
Completed. tested. Fixes last bugs.
